Fixed rest of @proavalon/proto implementation. - Circle CI runs correctly. - Dto's have all been moved into @proavalon/proto. - Set up testing for @proavalon/proto. - Moved backend integration test to root ProAvalon folder.
