+++ docker-compose-testing-circleci.yml
@@ -1,0 +1,20 @@
+version: '3'
+services:
+  redis-testing:
+    image: redis:5.0.7-alpine # Use stable version
+    container_name: redis-testing
+
+  backend-testing:
+    container_name: backend-testing
+    depends_on:
+      - redis-testing
+    build:
+      context: .
+      dockerfile: Dockerfile.test
+      args:
+        SERVICE: backend
+    environment:
+      SERVICE: backend
+    # volumes:
+    #   - ./packages/backend/src:/usr/src/app/packages/backend/src
+    #   - ./packages/proto/build:/usr/src/app/packages/proto/build
