+++ packages/backend/package.json
@@ -21,13 +21,7 @@
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
     "dev": "nest start --watch --preserveWatchOutput",
-    "test:int": "npm run redis-test:start && npm run redis-test:stop",
-    "test:int:circleci": "npm run redis-test:start:circleci && npm run redis-test:stop:circleci",
-    "test:int:jest": "jest --passWithNoTests --forceExit --runInBand --testRegex ispec\\.ts$",
-    "redis-test:start": "npm run redis-test:stop && echo Starting up redis integration testing server && docker-compose -f docker-compose-redis-testing.yml up --abort-on-container-exit --force-recreate",
-    "redis-test:start:circleci": "npm run redis-test:stop && echo Starting up redis integration testing server && docker-compose -f docker-compose-redis-testing-circleci.yml up --abort-on-container-exit",
-    "redis-test:stop": "echo Stopping redis integration testing server && docker-compose -f docker-compose-redis-testing.yml down -v",
-    "redis-test:stop:circleci": "echo Stopping redis integration testing server && docker-compose -f docker-compose-redis-testing-circleci.yml down -v"
+    "test:int:run": "jest --passWithNoTests --forceExit --runInBand --testRegex ispec\\.ts$"
   },
   "lint-staged": {
     "*.json": [
