--- packages/backend/src/forums/dto/create-forumpost.dto.ts
@@ -1,19 +1,0 @@
-import { IsDefined, IsNotEmpty } from 'class-validator';
-
-export class CreateForumPostDto {
-  @IsNotEmpty({
-    message: 'Title should not be empty.',
-  })
-  @IsDefined({
-    message: 'Title is missing.',
-  })
-  title!: string;
-
-  @IsNotEmpty({
-    message: 'Text should not be empty.',
-  })
-  @IsDefined({
-    message: 'Text is missing.',
-  })
-  text!: string;
-}
