+++ packages/backend/src/forums/forums.controller.ts
@@ -1,16 +1,13 @@
 import { Controller, Post, Get, Param, Body } from '@nestjs/common';
+import { CreateForumPostDto, CreateForumCommentDto } from '@proavalon/proto';
 import { ForumsService } from './forums.service';
-import { CreateForumPostDto } from './dto/create-forumpost.dto';
-import { CreateForumCommentDto } from './dto/create-forumcomment.dto';
 
 @Controller('forums')
 export class ForumsController {
   constructor(private readonly forumsService: ForumsService) {}
 
   @Post()
-  async addNewForumPost(
-    @Body() createForumPostDto: CreateForumPostDto,
-  ) {
+  async addNewForumPost(@Body() createForumPostDto: CreateForumPostDto) {
     const postId = await this.forumsService.addPost(createForumPostDto);
     return { id: postId };
   }
@@ -19,7 +16,9 @@ export class ForumsController {
   async addNewForumComment(
     @Body() createForumCommentDto: CreateForumCommentDto,
   ) {
-    const commentId = await this.forumsService.addComment(createForumCommentDto);
+    const commentId = await this.forumsService.addComment(
+      createForumCommentDto,
+    );
     return { id: commentId };
   }
