+++ packages/proto/src/index.ts
@@ -4,3 +4,5 @@ export {
 } from 'class-transformer-validator';
 
 export * from './proto/lobbyProto';
+export * from './proto/forumProto';
+export * from './proto/userProto';
