+++ packages/proto/src/proto/forumProto.ts
@@ -14,3 +14,21 @@ export class CreateForumCommentDto {
   })
   parentId!: string;
 }
+
+export class CreateForumPostDto {
+  @IsNotEmpty({
+    message: 'Title should not be empty.',
+  })
+  @IsDefined({
+    message: 'Title is missing.',
+  })
+  title!: string;
+
+  @IsNotEmpty({
+    message: 'Text should not be empty.',
+  })
+  @IsDefined({
+    message: 'Text is missing.',
+  })
+  text!: string;
+}
